% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjustr.R
\name{adjustr}
\alias{adjustr}
\title{NMR Number of Scans Estimator}
\usage{
adjustr(sig, noi, goal, ns)
}
\arguments{
\item{sig}{The signal}

\item{noi}{The noise}

\item{goal}{The target signal to noise}

\item{ns}{The number of scans used to create the provided signal and noise values}
}
\value{
The number of scans (rounded to the nearest multiple of 4 capped at 512) necessary to achieve the goal signal to noise
}
\description{
NMR Number of Scans Estimator
}
