% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cali.R
\name{cali}
\alias{cali}
\title{NMR spectral TSP calibration}
\usage{
cali(x, p)
}
\arguments{
\item{x}{The spectrum you want to calibrate as an array}

\item{p}{The matched ppm variable to the x you want to calibrate}
}
\value{
A spectrum that is calibrated to tsp
}
\description{
NMR spectral TSP calibration
}
