% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readin.R
\name{chemShift_}
\alias{chemShift_}
\title{Calculate chemical shift axis}
\usage{
chemShift_(swidth, offset, si)
}
\value{
num array, chemical shift for resp experiment
}
\description{
Calculate chemical shift axis
}
\section{}{
NA
}

\keyword{internal}
