% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crea.R
\name{crea}
\alias{crea}
\title{Creatinine Area Estimator}
\usage{
crea(x, p, n, c3 = c(3, 3.1), c4 = c(4, 4.1))
}
\arguments{
\item{x}{The spectrum of which you want to calculate the total area}

\item{n}{The noise estimation from the pproc}

\item{c3}{The ppm lower and upper limits defining the creatinine signal at ppm 3}

\item{c4}{The ppm lower and upper limits defining the creatinine signal at ppm 4}

\item{ppm}{The matched ppm variable to x}
}
\value{

}
\description{
Creatinine Area Estimator
}
