% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readin.R
\name{filterExp_files_}
\alias{filterExp_files_}
\title{Filter for experiment files}
\usage{
filterExp_files_(pars, exp_type, f_list, n_max)
}
\arguments{
\item{pars}{the data frame produced by .extract_pars1d}

\item{exp_type}{the experiment defined in the readin function}

\item{f_list}{the list produced by detect1d_procs}

\item{n_max}{the argument defined in the readin function}
}
\value{

}
\description{
Filter for experiment files
}
