% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flip_.R
\name{flip_}
\alias{flip_}
\title{Spectral Flipping}
\usage{
flip_(x, p, sh)
}
\arguments{
\item{x}{the spectra wanting to be flipped}

\item{p}{the match ppm variable}

\item{sh}{the chemical shift used to calculate orientation}
}
\value{
Returns a spectrum (array) with the correct orientation
}
\description{
Spectral Flipping
}
\author{
Kyle Bario \email{kylebario1@gmail.com}
}
