% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_idx.R
\name{get_idx}
\alias{get_idx}
\title{Get ppm indexes}
\usage{
get_idx(r, p)
}
\arguments{
\item{r}{the lower and upper bounds of the ppm region you wish to find the indexes of}

\item{p}{the ppm variable you want to find the indexes of}
}
\value{
An array containing the indexes between the lower and upper bounds
}
\description{
Get ppm indexes
}
\author{
Torben Kimhofer \email{torben.kimhofer@murdoch.edu.au}
}
