% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noi_.R
\name{noi_}
\alias{noi_}
\title{Noise Estimation}
\usage{
noi_(x, p)
}
\arguments{
\item{x}{The read in spectra}

\item{p}{the matched ppm var to your x}
}
\value{
an estimation of noise level
}
\description{
Noise Estimation
}
