% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pproc.R
\name{pproc}
\alias{pproc}
\title{Spectral Preprocessing}
\usage{
pproc(x, p)
}
\arguments{
\item{x}{The spectra that you are processing}

\item{p}{The ppm matched to your spectra}
}
\value{
This function returns:
\enumerate{
\item \strong{x}: a fully preprocessed spectrum ready for it's dilution to be estimated
\item \strong{p}: the matched ppm variable for the preprocessed x
\item \strong{o}: the original X spectra that was input into the function
\item \strong{n}: the estimated value of noise for this spectrum
}
}
\description{
Spectral Preprocessing
}
