% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readin.R
\name{readin}
\alias{readin}
\title{Import 1D NMR spectra}
\usage{
readin(
  path,
  exp_type = list(exp = c("PROF_URINE_NOESY")),
  n_max = 1000,
  filter = TRUE,
  recursive = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{path}{Given as a string, the path to the overarching files containing the NMR spectrum.}

\item{exp_type}{Defines the name of the experiment used to acquire the NMR. is given in the format: exp_type = list(exp = c("\emph{insert experiment name here}").}

\item{n_max}{Default = 1000}

\item{filter}{Default = TRUE}

\item{recursive}{Default = TRUE}

\item{verbose}{Default = TRUE}
}
\value{
The function exports the following three objects into the currently active R environment (no variable assignments needed):
\itemize{
\item X, num matrix:  NMR data, spectra in rows
\item ppm, num array - chemical shift positions, length matches to columns in X
\item meta, data.frame - spectrometer metadata as extracted from individual \emph{acqus} files, row-matched to X
}
Objects in the R environment with the same variable names will be overwritten.


}
\description{
Import 1D NMR spectra
}
\details{
This function imports TopSpin processed NMR spectra as well as spectrometer and processing parameters found in files \emph{acqus} and \emph{procs}. Experiments can be filtered according to data acquisition variables using the \code{exp_type} argument: For example, to read standard 1D NMR experiments use \code{exp_type=list(exp='noesygppr1d')}. More than one argument can be provided as list element.
}
\section{}{
NA
}

\examples{
path<-system.file('extdata/',  package = 'metabom8')
read1d(path,  exp_type=list(pulprog='noesygppr1d'))
}
\author{
Torben Kimhofer \email{torben.kimhofer@murdoch.edu.au}
}
