% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shift_pickr.R
\name{shift_pickr}
\alias{shift_pickr}
\title{Chemical Shift Picking}
\usage{
shift_pickr(x, p, sh, pm = 0.005)
}
\arguments{
\item{x}{The spectrum of which you want to calculate the total area}

\item{p}{The matched ppm variable to x}

\item{sh}{The ppm lower and upper limits defining the region you want to search in}

\item{pm}{The plus/minus value you want to add or subtract from the peak. Default = 0.005}
}
\value{
An array of values mapped to defined peak
}
\description{
Chemical Shift Picking
}
