% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citra.R
\name{citra}
\alias{citra}
\title{Citrate Area Estimation}
\usage{
citra(x, p, n, sh = c(5.2, 5.3))
}
\arguments{
\item{x}{The spectrum of which you want to calculate the total area}

\item{p}{The matched ppm variable to x}

\item{n}{The noise estimation from the pproc}

\item{sh}{The ppm lower and upper limits defining the cirtate signal}
}
\value{
An estimated value of the citrate area
}
\description{
Citrate Area Estimation
}
