% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pproc.R
\name{pproc}
\alias{pproc}
\title{Spectral Preprocessing}
\usage{
pproc(x, p)
}
\arguments{
\item{x}{The spectra that you are processing}

\item{p}{The ppm matched to your spectra}
}
\value{
This function returns a fully preprocessed spectra ready for it's dilution to be estimated
}
\description{
Spectral Preprocessing
}
